fixed merge conflicts with new myProfile page
